refactor&feat: use styled inputs & new creator structure
