+++ apps/web/src/pages/course/new.tsx
@@ -6,78 +6,31 @@ import { CourseTagsRecord } from '@studio/commons';
 import { getAuthTokenCookie } from '../../lib/cookieUtils';
 import { Header } from '../../components/header/header';
 import Button from '@studio/ui/components/interactivity/cta/button';
-import Form, {
+import {
   FormAreaTextInput,
-  FormBody,
   FormSelectMultipleInput,
   FormTextInput,
 } from '@studio/ui/components/interactivity/form';
 import { createCourse } from '../../contexts/courses/application/CreateCourse';
+import { CreatorHeader } from '../../components/creator/header';
+import CreateNewCourseForm from '../../components/creator/course/newCourseForm';
 
-export default function CreateNewCourseForm() {
+export default function CreateNewCourseView() {
   const router = useRouter();
 
   useEffect(() => {
     if (!getAuthTokenCookie()) router.push('/');
   }, [router]);
 
-  const [title, setTitle] = useState<string>();
-  const [tags, setTags] = useState<string[]>([]);
-  const [description, setDescription] = useState<string>();
-
-  const onSubmitCreateCourse = async (e: React.MouseEvent<HTMLElement>) => {
-    e.preventDefault();
-
-    if (!title) {
-      return;
-    }
-    if (!description) {
-      return;
-    }
-
-    await createCourse(
-      {
-        title,
-        tags,
-        description,
-      },
-      getAuthTokenCookie() || ''
-    );
-
-    router.push('/creator');
-  };
-
   return (
     <Fragment>
       <Header />
-      <Form>
-        <FormBody>
-          <FormTextInput
-            Name="Title"
-            placeholder="Title"
-            type="text"
-            onChange={(e) => setTitle(e.currentTarget.value)}
-          />
-          <FormAreaTextInput
-            Name="Description"
-            placeholder="Description"
-            type="text"
-            onChange={(e) => setDescription(e.currentTarget.value)}
-          />
-          <FormSelectMultipleInput
-            Name="Tags"
-            Values={CourseTagsRecord}
-            OnSelect={(values) => setTags(values)}
-          />
-
-          <Button
-            Label="Create Course!"
-            Size="Medium"
-            Type="Primary"
-            onClick={onSubmitCreateCourse}
-          />
-        </FormBody>
-      </Form>
+      <CreatorHeader title="Create new course" />
+      <div className="row">
+        <div className="column">
+          <CreateNewCourseForm />
+        </div>
+      </div>
     </Fragment>
   );
 }
